Refactor config structure
